split config files into separate files
