+++ project/libs/files-storage-lib/src/config/src/files-storage-config.module.ts
@@ -1,8 +1,6 @@
 import { Module } from '@nestjs/common';
 import { ConfigModule } from '@nestjs/config';
-import applicationConfig from '../../../../shared/core/src/config/app.config';
-import mongoConfig from '../../../../shared/core/src/config/mongo.config';
-import storageConfig from '../../../../shared/core/src/config/storage.config';
+import { applicationConfig, mongoConfig, storageConfig } from '@project/core';
 
 const ENV_FILES_STORAGE_FILE_PATH = 'apps/files-storage/.env';
