+++ project/libs/shared/core/src/config/app/app.constants.ts
@@ -1,5 +1,3 @@
 export const ENVIRONMENTS = ['development', 'production', 'stage'] as const;
 export const DEFAULT_PORT = 3000;
 export const DEFAULT_HOST = 'localhost';
-export const DEFAULT_MONGO_PORT = 27017;
-export const DEFAULT_MONGO_UI_PORT = 8081;
