+++ project/libs/shared/core/src/config/app/environment.type.ts
@@ -1,3 +1,3 @@
-import { ENVIRONMENTS } from './config.constants';
+import { ENVIRONMENTS } from './app.constants';
 
 export type Environment = (typeof ENVIRONMENTS)[number];
