+++ project/libs/shared/core/src/config/app/index.ts
@@ -1,0 +1,1 @@
+export * from './app.config';
