+++ project/libs/shared/core/src/config/index.ts
@@ -1,7 +1,7 @@
-export * from './app.config';
-export * from './mongo.config';
+export * from './app';
+export * from './mongo';
 export * from './utils';
 export * from './registers.enum';
-export * from './environment.type';
-export * from './storage.config';
-export * from './config.constants';
+export * from './app/environment.type';
+export * from './storage';
+export * from './app/app.constants';
