+++ project/libs/shared/core/src/config/mongo/index.ts
@@ -1,0 +1,3 @@
+export * from './mongo.constants';
+export * from './mongo.config';
+export * from './mongo.interface';
