+++ project/libs/shared/core/src/config/mongo/mongo.constants.ts
@@ -1,0 +1,3 @@
+export const MONGO_VALIDATION_ERROR = 'MongoDB Config Validation Error';
+export const DEFAULT_MONGO_PORT = 27017;
+export const DEFAULT_MONGO_UI_PORT = 8081;
