+++ project/libs/shared/core/src/config/mongo/mongo.interface.ts
@@ -1,0 +1,9 @@
+export interface MongoConfig {
+  host: string;
+  name: string;
+  port: number;
+  uiPort: number;
+  user: string;
+  password: string;
+  authBase: string;
+}
