+++ project/libs/shared/core/src/config/storage/index.ts
@@ -1,0 +1,3 @@
+export * from './storage.constants';
+export * from './storage.interface';
+export * from './storage.config';
