+++ project/libs/shared/core/src/config/storage/storage.constants.ts
@@ -1,0 +1,1 @@
+export const STORAGE_VALIDATION_ERROR = 'Uploads Storage Validation Error';
