+++ project/libs/shared/core/src/config/storage/storage.interface.ts
@@ -1,0 +1,3 @@
+export interface StorageConfig {
+  rootPath: string;
+}
